:truck: [#1206] 컴포넌트 모듈 변경 BaseFeature -> MyInfoFeature
