+++ Projects/App/Sources/Application/AppComponent+MyInfo.swift
@@ -34,4 +34,8 @@ extension AppComponent {
     var profilePopupFactory: any ProfilePopupFactory {
         ProfilePopupComponent(parent: self)
     }
+    
+    var togglePopupFactory: any TogglePopupFactory {
+        TogglePopupComponent(parent: self)
+    }
 }
