+++ Projects/Features/MyInfoFeature/Sources/Components/TogglePopupComponent.swift
@@ -1,36 +1,27 @@
+import MyInfoFeatureInterface
+import NeedleFoundation
 import UIKit
 
-public protocol TogglePopupFactory {
-    func makeView(
-        titleString: String,
-        firstItemString: String,
-        secondItemString: String,
-        cancelButtonText: String,
-        confirmButtonText: String,
-        descriptionText: String,
-        completion: (() -> Void)?,
-        cancelCompletion: (() -> Void)?
-    ) -> UIViewController
-}
-
-public extension TogglePopupFactory {
-    func makeView(
+public final class TogglePopupComponent: Component<EmptyDependency>, TogglePopupFactory {
+    public func makeView(
         titleString: String,
         firstItemString: String,
         secondItemString: String,
         cancelButtonText: String = "취소",
         confirmButtonText: String = "확인",
-        descriptionText: String = "",
+        firstDescriptionText: String = "",
+        secondDescriptionText: String = "",
         completion: (() -> Void)? = nil,
         cancelCompletion: (() -> Void)? = nil
     ) -> UIViewController {
-        self.makeView(
+        return TogglePopupViewController(
             titleString: titleString,
             firstItemString: firstItemString,
             secondItemString: secondItemString,
             cancelButtonText: cancelButtonText,
             confirmButtonText: confirmButtonText,
-            descriptionText: descriptionText,
+            firstDescriptionText: firstDescriptionText,
+            secondDescriptionText: secondDescriptionText,
             completion: completion,
             cancelCompletion: cancelCompletion
         )
